requirements.txt
# Add your pinned requirements so that they can be easily installed with:
# pip install -r requirements.txt
# Remember to also add them in setup.cfg but unpinned.
# Example:
# numpy==1.13.3
# scipy==1.0

Flask
Flask-HTTPAuth
Flask-RESTful
flask-restful-swagger-2
Flask-Cors
boto3
docutils
numpy
redis
requests
rq
uWSGI
fluent-logger
python-magic
Sphinx
google-cloud
ply
pycurl
passlib
urlparse3
